refactor: Use react-redux connect() to access the store in the Vehicles component
This pattern abstracts away having to deal with contextTypes and context directly, by generating a
container component with the appropriate contextTypes and render/lifecycle methods necessary for
using a Redux store. The generated mount/unmount lifecycle methods will subscribe/unsubscribe to the
Redux store, forcing updates on changes, and render will call the child component passed to the
connect() function.

Closes #2