refactor: Remove explicit/manual import and setup of chai assertions
The karma-chai plugin, installed and configured in the last commit, injects the necessary methods
from chai into all test files, so there is no longer a need to repeat the same import statement in
each file. The same goes for importing sinon, because the karma-sinon plugin was similarily setup
during the last commit.